+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+*.*.* / 2023-2-**
+------------------
+* export a Fastify Module
+* https/http agents options
+
 2.7.0 / 2022-07-09
 ------------------
 * accept XHTML